fix(Subs): distinguish visual block mode and visual line mode
Until now, it was only working properly for visual block mode. If visual
line had a short last line selected, it wouldn't work because in visual
block mode, this means that the range is as short as the last line.

So if the last line was a line with 0 chars, which can happen easily
in coding files, it would not work.

Example:

```typescript
const foo = "bar";
const bar = "bar";

console.log(foo);
````

In this example, selecting the first two lines and doing a `Subs` will
work as expected. However, if the third line is selected, nothing will
happen because the third line has 0 chars and hence the visual block
mode will have a range of 0 chars.

**This change fixes that.**