refactor: start_replacing_command_with_first_part to any part
This way, `start_replacing_command_with_first_part` can be done by being
on the first part of the word and calling `start_replacing_command_with_part(1)`.
But it's more powerful than that. You can use `start_replacing_command_with_part(2)`
to replace the current part and the next part of the word. And so on ...